# backend/internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/health": {
            "get": {
                "description": "to check http server health",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "show http server health",
                "responses": {}
            }
        },
        "/words/add": {
            "post": {
                "description": "Add word",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "word"
                ],
                "summary": "Add word",
                "parameters": [
                    {
                        "description": "AddWordRequest body",
                        "name": "AddWordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddWordRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.AddWordRequest": {
            "type": "object",
            "properties": {
                "definitions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Definition"
                    }
                },
                "user_id": {
                    "type": "string"
                },
                "word": {
                    "$ref": "#/definitions/models.Word"
                }
            }
        },
        "models.Definition": {
            "type": "object",
            "properties": {
                "definition": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                }
            }
        },
        "models.Word": {
            "type": "object",
            "properties": {
                "example": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        }
    }
}